fix: improved code readability
