+++ pallets/service-request/src/lib.rs
@@ -1,20 +1,16 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
+pub use pallet::*;
+
 use frame_support::{
 	codec::{Decode, Encode},
 	dispatch::DispatchResultWithPostInfo,
 	pallet_prelude::*,
-	scale_info::TypeInfo,
-	sp_runtime::{
-		traits::{AccountIdConversion, Hash, Zero},
-		RuntimeDebug,
-	},
+	sp_runtime::{traits::Hash, RuntimeDebug},
 	sp_std::prelude::*,
-	traits::{Currency, ExistenceRequirement, UnixTime},
-	PalletId,
+	traits::{Currency, UnixTime},
 };
 use frame_system::pallet_prelude::*;
-pub use pallet::*;
 use primitives_verification_status::VerificationStatusTrait;
 use traits_labs::LabsProvider;
 
@@ -24,154 +20,19 @@ mod mock;
 #[cfg(test)]
 mod tests;
 
+pub mod functions;
+pub mod impl_service_request;
 pub mod interface;
-pub use interface::SeviceRequestInterface;
-
+pub mod types;
 pub mod weights;
+
+pub use interface::SeviceRequestInterface;
+pub use types::*;
 pub use weights::WeightInfo;
-
-#[derive(Encode, Decode, Clone, RuntimeDebug, PartialEq, Eq, TypeInfo)]
-pub enum RequestStatus {
-	Open,
-	WaitingForUnstaked,
-	Unstaked,
-	Claimed,
-	Processed,
-	Finalized,
-}
-impl Default for RequestStatus {
-	fn default() -> Self {
-		RequestStatus::Open
-	}
-}
-
-#[derive(Clone, Decode, Default, Encode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct Request<AccountId, Balance, Hash> {
-	pub hash: Hash,
-	pub requester_address: AccountId,
-	pub lab_address: Option<AccountId>,
-	pub country: Vec<u8>,
-	pub region: Vec<u8>,
-	pub city: Vec<u8>,
-	pub service_category: Vec<u8>,
-	pub staking_amount: Balance,
-	pub status: RequestStatus,
-	pub created_at: u128,
-	pub updated_at: Option<u128>,
-	pub unstaked_at: Option<u128>,
-}
-#[allow(clippy::too_many_arguments)]
-impl<AccountId, Balance, Hash> Request<AccountId, Balance, Hash> {
-	pub fn new(
-		hash: Hash,
-		requester_address: AccountId,
-		country: Vec<u8>,
-		region: Vec<u8>,
-		city: Vec<u8>,
-		service_category: Vec<u8>,
-		staking_amount: Balance,
-		created_at: u128,
-	) -> Self {
-		Self {
-			hash,
-			requester_address,
-			lab_address: None,
-			country,
-			region,
-			city,
-			service_category,
-			staking_amount,
-			status: RequestStatus::default(),
-			created_at,
-			updated_at: None,
-			unstaked_at: None,
-		}
-	}
-}
-
-#[derive(Clone, Decode, Encode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct ServiceOffer<AccountId, Balance, Hash> {
-	pub request_hash: Hash,
-	pub lab_address: AccountId,
-	pub service_id: Hash,
-	pub testing_price: Balance,
-	pub qc_price: Balance,
-}
-impl<AccountId, Balance, Hash> ServiceOffer<AccountId, Balance, Hash> {
-	pub fn new(
-		request_hash: Hash,
-		lab_address: AccountId,
-		service_id: Hash,
-		testing_price: Balance,
-		qc_price: Balance,
-	) -> Self {
-		Self { request_hash, lab_address, service_id, testing_price, qc_price }
-	}
-}
-
-#[derive(Clone, Decode, Encode, Eq, PartialEq, RuntimeDebug, TypeInfo)]
-pub struct ServiceInvoice<AccountId, Balance, Hash> {
-	pub request_hash: Hash,
-	pub order_id: Hash,
-	pub service_id: Hash,
-	pub customer_address: AccountId,
-	pub seller_address: AccountId,
-	pub dna_sample_tracking_id: Vec<u8>,
-	pub testing_price: Balance,
-	pub qc_price: Balance,
-	pub pay_amount: Balance,
-}
-#[allow(clippy::too_many_arguments)]
-impl<AccountId, Balance, Hash> ServiceInvoice<AccountId, Balance, Hash> {
-	pub fn new(
-		request_hash: Hash,
-		order_id: Hash,
-		service_id: Hash,
-		customer_address: AccountId,
-		seller_address: AccountId,
-		dna_sample_tracking_id: Vec<u8>,
-		testing_price: Balance,
-		qc_price: Balance,
-		pay_amount: Balance,
-	) -> Self {
-		Self {
-			request_hash,
-			order_id,
-			service_id,
-			customer_address,
-			seller_address,
-			dna_sample_tracking_id,
-			testing_price,
-			qc_price,
-			pay_amount,
-		}
-	}
-}
 
 #[frame_support::pallet]
 pub mod pallet {
-	use crate::*;
-
-	pub const PALLET_ID: PalletId = PalletId(*b"reqsrvc!");
-
-	pub type AccountIdOf<T> = <T as frame_system::Config>::AccountId;
-	pub type CurrencyOf<T> = <T as self::Config>::Currency;
-	pub type BalanceOf<T> = <CurrencyOf<T> as Currency<AccountIdOf<T>>>::Balance;
-	pub type HashOf<T> = <T as frame_system::Config>::Hash;
-	pub type AdminOf<T> = AccountIdOf<T>;
-	pub type RequestOf<T> = Request<AccountIdOf<T>, BalanceOf<T>, HashOf<T>>;
-	pub type ServiceOfferOf<T> = ServiceOffer<AccountIdOf<T>, BalanceOf<T>, HashOf<T>>;
-	pub type ServiceInvoiceOf<T> = ServiceInvoice<AccountIdOf<T>, BalanceOf<T>, HashOf<T>>;
-	pub type RequestIdOf<T> = HashOf<T>;
-	pub type RequesterIdOf<T> = AccountIdOf<T>;
-	pub type LabIdOf<T> = AccountIdOf<T>;
-	pub type CountryOf = Vec<u8>;
-	pub type RegionOf = Vec<u8>;
-	pub type CityOf = Vec<u8>;
-	pub type ServiceCategoryOf = Vec<u8>;
-	pub type ServiceIdOf<T> = HashOf<T>;
-	pub type OrderIdOf<T> = HashOf<T>;
-	pub type DNASampleTrackingIdOf = Vec<u8>;
+	use super::*;
 
 	#[pallet::config]
 	pub trait Config: frame_system::Config {
@@ -474,548 +335,5 @@ pub mod pallet {
 				Err(error) => Err(error.into()),
 			}
 		}
-	}
-}
-
-/// Pallet Methods
-impl<T: Config> Pallet<T> {
-	pub fn staking_account_id(request_id: ServiceIdOf<T>) -> AccountIdOf<T> {
-		PALLET_ID.into_sub_account(request_id)
-	}
-}
-
-/// Service Request Interface Implementation
-impl<T: Config> SeviceRequestInterface<T> for Pallet<T> {
-	type Error = Error<T>;
-	type Balance = BalanceOf<T>;
-	type Request = RequestOf<T>;
-	type Admin = AdminOf<T>;
-	type ServiceOffer = ServiceOfferOf<T>;
-	type ServiceInvoice = ServiceInvoiceOf<T>;
-	type RequestId = RequestIdOf<T>;
-	type RequesterId = RequesterIdOf<T>;
-	type LabId = LabIdOf<T>;
-	type Country = CountryOf;
-	type Region = RegionOf;
-	type City = CityOf;
-	type ServiceCategory = ServiceCategoryOf;
-	type ServiceId = ServiceIdOf<T>;
-	type OrderId = OrderIdOf<T>;
-	type DNASampleTrackingId = DNASampleTrackingIdOf;
-
-	fn generate_request_id(
-		requester_id: Self::RequesterId,
-		country: Self::Country,
-		region: Self::Region,
-		city: Self::City,
-		service_category: Self::ServiceCategory,
-	) -> Self::RequestId {
-		let mut seed = requester_id.encode();
-		let account_info = frame_system::Pallet::<T>::account(requester_id);
-
-		seed.append(&mut account_info.nonce.encode());
-		seed.append(&mut country.encode());
-		seed.append(&mut region.encode());
-		seed.append(&mut city.encode());
-		seed.append(&mut service_category.encode());
-
-		T::Hashing::hash(&seed)
-	}
-
-	fn create_request(
-		requester_id: Self::RequesterId,
-		country: Self::Country,
-		region: Self::Region,
-		city: Self::City,
-		service_category: Self::ServiceCategory,
-		staking_amount: Self::Balance,
-	) -> Result<Self::Request, Self::Error> {
-		if staking_amount.is_zero() {
-			return Err(Error::<T>::NotValidAmount)
-		}
-
-		let request_id = Self::generate_request_id(
-			requester_id.clone(),
-			country.clone(),
-			region.clone(),
-			city.clone(),
-			service_category.clone(),
-		);
-
-		let now = T::TimeProvider::now().as_millis();
-
-		match CurrencyOf::<T>::transfer(
-			&requester_id,
-			&Self::staking_account_id(request_id),
-			staking_amount,
-			ExistenceRequirement::KeepAlive,
-		) {
-			Ok(_) => {
-				let request = Request::new(
-					request_id,
-					requester_id.clone(),
-					country.clone(),
-					region.clone(),
-					city.clone(),
-					service_category.clone(),
-					staking_amount,
-					now,
-				);
-
-				let mut request_ids = RequestByAccountId::<T>::get(requester_id.clone());
-				request_ids.push(request_id);
-
-				StakingAccountIdByRequestId::<T>::insert(
-					request_id,
-					Self::staking_account_id(request_id),
-				);
-				RequestByAccountId::<T>::insert(requester_id.clone(), request_ids);
-				RequestById::<T>::insert(request_id, request.clone());
-
-				let service_count = ServiceCountRequest::<T>::get((
-					country.clone(),
-					region.clone(),
-					city.clone(),
-					service_category.clone(),
-				));
-
-				ServiceCountRequest::<T>::insert(
-					(country, region, city, service_category),
-					service_count.wrapping_add(1),
-				);
-
-				Ok(request)
-			},
-			Err(dispatch) => match dispatch {
-				sp_runtime::DispatchError::Other(_) => Err(Error::<T>::Other),
-				sp_runtime::DispatchError::CannotLookup => Err(Error::<T>::CannotLookup),
-				sp_runtime::DispatchError::BadOrigin => Err(Error::<T>::BadOrigin),
-				sp_runtime::DispatchError::TooManyConsumers => Err(Error::<T>::TooManyConsumers),
-				sp_runtime::DispatchError::ConsumerRemaining => Err(Error::<T>::ConsumerRemaining),
-				sp_runtime::DispatchError::NoProviders => Err(Error::<T>::NoProviders),
-				sp_runtime::DispatchError::Token(_) => Err(Error::<T>::Token),
-				sp_runtime::DispatchError::Arithmetic(_) => Err(Error::<T>::Arithmetic),
-				sp_runtime::DispatchError::Module(_) => Err(Error::<T>::Arithmetic),
-			},
-		}
-	}
-
-	fn unstake(
-		requester_id: Self::RequesterId,
-		request_id: Self::RequestId,
-	) -> Result<Self::Request, Self::Error> {
-		let request = RequestById::<T>::get(request_id);
-
-		if request.is_none() {
-			return Err(Error::<T>::RequestNotFound)
-		}
-
-		let mut request = request.unwrap();
-
-		if request.requester_address != requester_id {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		if request.status != RequestStatus::Open {
-			return Err(Error::<T>::RequestUnableToUnstake)
-		}
-
-		request.status = RequestStatus::WaitingForUnstaked;
-		let now: u128 = T::TimeProvider::now().as_millis();
-		request.unstaked_at = Some(now);
-
-		RequestById::<T>::insert(request_id, request.clone());
-
-		Ok(request)
-	}
-
-	fn retrieve_unstaked_amount(
-		admin: Self::Admin,
-		request_id: Self::RequestId,
-	) -> Result<Self::Request, Self::Error> {
-		if admin != AdminKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		let request = RequestById::<T>::get(request_id);
-
-		if request.is_none() {
-			return Err(Error::<T>::RequestNotFound)
-		}
-
-		let mut request = request.unwrap();
-
-		if request.status != RequestStatus::WaitingForUnstaked {
-			return Err(Error::<T>::RequestUnableToRetrieveUnstake)
-		}
-
-		let requester_id = request.requester_address.clone();
-
-		let now: u128 = T::TimeProvider::now().as_millis();
-		let unstaked_at: u128 = request.unstaked_at.unwrap();
-		// TODO: move to constant runtime config
-		let six_days: u128 = 3600_u128 * 144_u128 * 1000_u128;
-
-		if (now - unstaked_at) == six_days {
-			return Err(Error::<T>::RequestWaitingForUnstaked)
-		}
-
-		match CurrencyOf::<T>::transfer(
-			&Self::staking_account_id(request_id),
-			&requester_id,
-			request.staking_amount,
-			ExistenceRequirement::AllowDeath,
-		) {
-			Ok(_) => {
-				Self::deposit_event(Event::StakingAmountRefunded(
-					requester_id,
-					request_id,
-					request.staking_amount,
-				));
-
-				request.status = RequestStatus::Unstaked;
-
-				RequestById::<T>::insert(request_id, request.clone());
-
-				Ok(request)
-			},
-			Err(dispatch) => match dispatch {
-				sp_runtime::DispatchError::Other(_) => Err(Error::<T>::Other),
-				sp_runtime::DispatchError::CannotLookup => Err(Error::<T>::CannotLookup),
-				sp_runtime::DispatchError::BadOrigin => Err(Error::<T>::BadOrigin),
-				sp_runtime::DispatchError::TooManyConsumers => Err(Error::<T>::TooManyConsumers),
-				sp_runtime::DispatchError::ConsumerRemaining => Err(Error::<T>::ConsumerRemaining),
-				sp_runtime::DispatchError::NoProviders => Err(Error::<T>::NoProviders),
-				sp_runtime::DispatchError::Token(_) => Err(Error::<T>::Token),
-				sp_runtime::DispatchError::Arithmetic(_) => Err(Error::<T>::Arithmetic),
-				sp_runtime::DispatchError::Module(_) => Err(Error::<T>::Arithmetic),
-			},
-		}
-	}
-
-	fn claim_request(
-		lab_id: Self::LabId,
-		request_id: Self::RequestId,
-		service_id: Self::ServiceId,
-		testing_price: Self::Balance,
-		qc_price: Self::Balance,
-	) -> Result<(Self::Request, Self::ServiceOffer), Self::Error> {
-		let request = RequestById::<T>::get(request_id);
-
-		if request.is_none() {
-			return Err(Error::<T>::RequestNotFound)
-		}
-
-		let mut request = request.unwrap();
-
-		if request.status == RequestStatus::WaitingForUnstaked ||
-			request.status == RequestStatus::Unstaked
-		{
-			return Err(Error::<T>::RequestAlreadyUnstaked)
-		}
-
-		if request.status == RequestStatus::Claimed {
-			return Err(Error::<T>::RequestAlreadyClaimed)
-		}
-
-		if request.status != RequestStatus::Open {
-			return Err(Error::<T>::RequestUnableToClaimed)
-		}
-
-		let lab_status = T::Labs::lab_verification_status(&lab_id);
-
-		if lab_status.is_none() {
-			return Err(Error::<T>::LabNotFound)
-		}
-
-		let lab_status = lab_status.unwrap();
-
-		let service_offer =
-			ServiceOffer::new(request_id, lab_id.clone(), service_id, testing_price, qc_price);
-
-		let now = T::TimeProvider::now().as_millis();
-
-		if lab_status.is_verified() {
-			request.status = RequestStatus::Claimed;
-			request.lab_address = Some(lab_id);
-			request.updated_at = Some(now);
-
-			RequestById::<T>::insert(request_id, request.clone());
-			ServiceOfferById::<T>::insert(request_id, service_offer.clone());
-		} else {
-			let mut request_ids = RequestsByLabId::<T>::get(lab_id.clone());
-			let found = request_ids.iter().find(|x| x == &&request_id);
-
-			if found.is_some() {
-				return Err(Error::<T>::RequestAlreadyInList)
-			}
-
-			request_ids.push(request_id);
-
-			RequestsByLabId::<T>::insert(lab_id, request_ids);
-		}
-
-		Ok((request, service_offer))
-	}
-
-	fn process_request(
-		requester_id: Self::RequesterId,
-		lab_id: Self::LabId,
-		request_id: Self::RequestId,
-		order_id: Self::OrderId,
-		dna_sample_tracking_id: Self::DNASampleTrackingId,
-		additional_staking_amount: Self::Balance,
-	) -> Result<Self::ServiceInvoice, Self::Error> {
-		let request = RequestById::<T>::get(request_id);
-
-		if request.is_none() {
-			return Err(Error::<T>::RequestNotFound)
-		}
-
-		let mut request = request.unwrap();
-
-		if requester_id != request.requester_address {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		if request.status == RequestStatus::WaitingForUnstaked ||
-			request.status == RequestStatus::Unstaked
-		{
-			return Err(Error::<T>::RequestAlreadyUnstaked)
-		}
-
-		if request.lab_address.is_none() {
-			return Err(Error::<T>::LabNotFound)
-		}
-
-		if request.status != RequestStatus::Claimed {
-			return Err(Error::<T>::RequestUnableToProccess)
-		}
-
-		if request.lab_address.clone().unwrap() != lab_id {
-			return Err(Error::<T>::LabNotFound)
-		}
-
-		let service_offer = ServiceOfferById::<T>::get(request_id).unwrap();
-		let pay_amount = request.staking_amount;
-		let testing_price = service_offer.testing_price;
-		let qc_price = service_offer.qc_price;
-		let total_price = testing_price + qc_price;
-		let mut excess = Zero::zero();
-		let mut final_pay_amount = pay_amount;
-
-		if pay_amount > total_price {
-			excess = pay_amount - total_price;
-			final_pay_amount = total_price;
-		}
-
-		if !excess.is_zero() {
-			match CurrencyOf::<T>::transfer(
-				&Self::staking_account_id(request_id),
-				&requester_id,
-				excess,
-				ExistenceRequirement::KeepAlive,
-			) {
-				Ok(_) => {
-					Self::deposit_event(Event::StakingAmountExcessRefunded(
-						requester_id.clone(),
-						request_id,
-						excess,
-					));
-				},
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		} else {
-			final_pay_amount = request.staking_amount + additional_staking_amount;
-
-			if final_pay_amount != total_price {
-				return Err(Error::<T>::NotValidAmount)
-			} else {
-				match CurrencyOf::<T>::transfer(
-					&requester_id,
-					&Self::staking_account_id(request_id),
-					additional_staking_amount,
-					ExistenceRequirement::KeepAlive,
-				) {
-					Ok(_) => {
-						Self::deposit_event(Event::StakingAmountIncreased(
-							requester_id.clone(),
-							request_id,
-							additional_staking_amount,
-						));
-					},
-					Err(dispatch) => match dispatch {
-						sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-						sp_runtime::DispatchError::CannotLookup =>
-							return Err(Error::<T>::CannotLookup),
-						sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-						sp_runtime::DispatchError::TooManyConsumers =>
-							return Err(Error::<T>::TooManyConsumers),
-						sp_runtime::DispatchError::ConsumerRemaining =>
-							return Err(Error::<T>::ConsumerRemaining),
-						sp_runtime::DispatchError::NoProviders =>
-							return Err(Error::<T>::NoProviders),
-						sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-						sp_runtime::DispatchError::Arithmetic(_) =>
-							return Err(Error::<T>::Arithmetic),
-						sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-					},
-				}
-			}
-		}
-
-		let service_invoice = ServiceInvoice::new(
-			request_id,
-			order_id,
-			service_offer.service_id,
-			requester_id,
-			lab_id,
-			dna_sample_tracking_id,
-			testing_price,
-			qc_price,
-			final_pay_amount,
-		);
-
-		let now = T::TimeProvider::now().as_millis();
-
-		ServiceInvoiceById::<T>::insert(request_id, service_invoice.clone());
-		ServiceInvoiceByOrderId::<T>::insert(order_id, service_invoice.clone());
-
-		request.status = RequestStatus::Processed;
-		request.updated_at = Some(now);
-		RequestById::<T>::insert(request_id, request);
-
-		Ok(service_invoice)
-	}
-
-	fn finalize_request(
-		admin: Self::Admin,
-		request_id: Self::RequestId,
-		test_result_success: bool,
-	) -> Result<Self::ServiceInvoice, Self::Error> {
-		if admin != AdminKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		let service_invoice = ServiceInvoiceById::<T>::get(request_id);
-
-		if service_invoice.is_none() {
-			return Err(Error::<T>::ServiceInvoiceNotFound)
-		}
-
-		let service_invoice = service_invoice.unwrap();
-
-		let mut request = RequestById::<T>::get(request_id).unwrap();
-
-		if request.status != RequestStatus::Processed {
-			return Err(Error::<T>::RequestUnableToFinalize)
-		}
-
-		let requester_id = service_invoice.customer_address.clone();
-		let lab_id = service_invoice.seller_address.clone();
-
-		let mut pay_amount = service_invoice.pay_amount;
-
-		if !test_result_success {
-			pay_amount = service_invoice.qc_price;
-
-			let testing_price = service_invoice.testing_price;
-
-			// Transfer testing price back to customer
-			match CurrencyOf::<T>::transfer(
-				&Self::staking_account_id(request_id),
-				&requester_id,
-				testing_price,
-				ExistenceRequirement::AllowDeath,
-			) {
-				Ok(_) => (),
-				Err(dispatch) => match dispatch {
-					sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-					sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-					sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-					sp_runtime::DispatchError::TooManyConsumers =>
-						return Err(Error::<T>::TooManyConsumers),
-					sp_runtime::DispatchError::ConsumerRemaining =>
-						return Err(Error::<T>::ConsumerRemaining),
-					sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-					sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-					sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-					sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-				},
-			}
-		}
-
-		// Transfer to lab_id
-		match CurrencyOf::<T>::transfer(
-			&Self::staking_account_id(request_id),
-			&lab_id,
-			pay_amount,
-			ExistenceRequirement::AllowDeath,
-		) {
-			Ok(_) => (),
-			Err(dispatch) => match dispatch {
-				sp_runtime::DispatchError::Other(_) => return Err(Error::<T>::Other),
-				sp_runtime::DispatchError::CannotLookup => return Err(Error::<T>::CannotLookup),
-				sp_runtime::DispatchError::BadOrigin => return Err(Error::<T>::BadOrigin),
-				sp_runtime::DispatchError::TooManyConsumers =>
-					return Err(Error::<T>::TooManyConsumers),
-				sp_runtime::DispatchError::ConsumerRemaining =>
-					return Err(Error::<T>::ConsumerRemaining),
-				sp_runtime::DispatchError::NoProviders => return Err(Error::<T>::NoProviders),
-				sp_runtime::DispatchError::Token(_) => return Err(Error::<T>::Token),
-				sp_runtime::DispatchError::Arithmetic(_) => return Err(Error::<T>::Arithmetic),
-				sp_runtime::DispatchError::Module(_) => return Err(Error::<T>::Arithmetic),
-			},
-		}
-
-		let now = T::TimeProvider::now().as_millis();
-
-		// Removed from customer request list
-		let mut request_ids = RequestByAccountId::<T>::get(requester_id.clone());
-		request_ids.retain(|&x| x != request_id);
-		RequestByAccountId::<T>::insert(requester_id, request_ids);
-
-		// Update service count request
-		let service_count_request = ServiceCountRequest::<T>::get((
-			&request.country,
-			&request.region,
-			&request.city,
-			&request.service_category,
-		));
-		ServiceCountRequest::<T>::insert(
-			(&request.country, &request.region, &request.city, &request.service_category),
-			service_count_request.wrapping_sub(1),
-		);
-
-		request.status = RequestStatus::Finalized;
-		request.updated_at = Some(now);
-		RequestById::<T>::insert(request_id, request);
-
-		Ok(service_invoice)
-	}
-
-	fn update_admin_key(
-		account_id: &T::AccountId,
-		admin_key: &T::AccountId,
-	) -> Result<(), Self::Error> {
-		if account_id.clone() != AdminKey::<T>::get().unwrap() {
-			return Err(Error::<T>::Unauthorized)
-		}
-
-		AdminKey::<T>::put(admin_key);
-
-		Ok(())
 	}
 }
